requirements.txt
# Обязательные для задания
rich>=13.7.0

# Опционально: БД (database/)
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
alembic>=1.13.0

# Для разработки
pytest>=7.4.0
black>=23.0.0
